+++ examples/components.html
@@ -24,6 +24,8 @@
       importShim("/dist/index.esm.js").then(({ ThreeElement }) => {
         class RotatingThingy extends ThreeElement.for(THREE.Mesh) {
           readyCallback() {
+            super.readyCallback()
+
             this.speed = this.getAttribute("speed")
 
             const geometry = document.createElement("three-dodecahedron-buffer-geometry")
